+++ resources/utils/serialize.py
@@ -1,0 +1,18 @@
+import json
+from datetime import datetime, date
+from time import struct_time, mktime
+import decimal
+
+# class for serializer
+class clsCustomJSONEncoder(json.JSONEncoder):
+    def default(self, o):
+        if isinstance(o, datetime):
+            return str(o)
+        if isinstance(o, date):
+            return str(o)
+        if isinstance(o, decimal.Decimal):
+            return float(o)
+        if isinstance(o, struct_time):
+            return datetime.fromtimestamp(mktime(o))
+        # Any other serializer if needed
+        return super(clsCustomJSONEncoder, self).default(o)
